Revert "Feature/marker visibility"
