+++ app/src/main/java/com/example/osmexample/MarkerType.java
@@ -1,0 +1,8 @@
+package com.example.osmexample;
+
+public enum MarkerType {
+    DEFAULT,
+    FISH,
+    MUSHROOM,
+    WALK
+}
