--- app/src/main/java/com/example/osmexample/model/ObjectType.java
@@ -1,8 +1,0 @@
-package com.example.osmexample.model;
-
-public enum ObjectType {
-    DEFAULT,
-    FISH,
-    MUSHROOM,
-    WALK
-}
