--- app/src/main/res/drawable/bg_orange.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#FFE082"/>
-    <stroke android:width="4dp" android:color="#FF6F00" />
-    <corners android:radius="10dp"/>
-    <padding android:left="0dp" android:top="0dp" android:right="0dp" android:bottom="0dp" />
-</shape>
