--- app/src/main/res/drawable/bg_white.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="#FFFFFF"/>
-    <stroke android:width="2dp" android:color="#000000" />
-    <padding android:left="2dp" android:top="2dp" android:right="2dp" android:bottom="2dp" />
-</shape>
